Crate raft

This is the Raft Distributed Consensus Protocol implemented for Rust. Raft is described as:

Raft is a consensus algorithm that is designed to be easy to understand. It's equivalent to Paxos in fault-tolerance and performance. The difference is that it's decomposed into relatively independent subproblems, and it cleanly addresses all major pieces needed for practical systems.

This implementation utilizes Cap'n Proto for its RPC, mio for it's async event loop.

If this package fails to build for you it is possibly because you do not have the capnp utility installed. You should be able to find appropriate packages for most popular distributions.

Consuming this library

Consuming this library works in a few parts:

Implement Store and StateMachine such that they will hook into your application.

Create a Raft with those impls which will spawn it's own Server and join with a cluster.

Interact with the cluster by issuing append() calls.

React to calls to apply() from the implemented StateMachine

It's important to note that issuing an append() call to Raft does not (at this time) necessarily gaurantee that the entry has been applied to the StateMachine, since it must be successfully replicated across a majority of nodes before it is applied.

State

// TODO

StateMachine

// TODO

Client Requests

// TODO

Modules

state_machine
The state machine which Raft applied commited entries too.

store
The persistent storage of Raft state.

Structs

LogIndex
The index of a log entry.

Raft
This is the primary interface with a Server in the cluster.

Term
The term of a log entry.

Enums

Error
raft::Errors are the composed variety of errors that can originate from the various libraries. With the exception of the Raft variant these are generated from try!() macros invoking on io::Error or capnp::Error by using FromError.

ErrorKind
Currently, this can only be:

Type Definitions

Result