Enum raft::Error

pub enum Error {
    CapnProto(Error),
    SchemaError(NotInSchema),
    Io(Error),
    Raft(ErrorKind),
}

raft::Errors are the composed variety of errors that can originate from the various libraries. With the exception of the Raft variant these are generated from try!() macros invoking on io::Error or capnp::Error by using FromError.

Variants

CapnProto

SchemaError

Io

Raft