Enum raft::ErrorKind

pub enum ErrorKind {
    RelatedNodeDown,
    CannotProceed,
    NotInCluster,
    BadResponse,
}

Currently, this can only be:

RelatedNodeDown - When the related Server is known to be down.

CannotProceed - When the related Server cannot proceed due to more than a majority of nodes being unavailable. TODO: Hook these up.

Variants

RelatedNodeDown

CannotProceed

NotInCluster

BadResponse