Enum raft::RaftError

pub enum RaftError {
    ConnectionLimitReached,
    InvalidClientId,
    ClusterViolation,
    UnknownConnectionType,
    InvalidPeerSet,
    ConnectionRegisterFailed,
    LeaderSearchExhausted,
}

A Raft Error represents a Raft specific error that consuming code is expected to handle gracefully.

Variants
ConnectionLimitReached

The server ran out of slots in the slab for new connections

InvalidClientId

A client reported an invalid client id

ClusterViolation

A consensus module reported back a leader not in the cluster.

UnknownConnectionType

A remote connection attempted to use an unknown connection type in the connection preamble

InvalidPeerSet

An invalid peer in in the peer set. Returned Server::new().

ConnectionRegisterFailed

Registering a connection failed

LeaderSearchExhausted

Failed to find a leader in the cluster. Try again later.