Module raft::interchange

Structs

Accepted

AppendEntries

AppendRequest

IndexRange

Rejected

RequestVote

Enums

ClientRequest
Data interchange request format for Client <-> Node Communication. Each variant of this is a command which can be asked of the RaftNode after it is spun up with node.spinup() The node attached to this application will poll it's channel regularly and return results on the channel. If you're wondering where vote requesting is, it's hidden within the module. TODO: Currently requests are not queued or gaurenteed to serviced in order. This is probably a bad thing as most clients will .send() then .recv(). We can probably make a service queue for this.

RemoteProcedureCall
Data interchange format for RPC calls. These should match directly to the Raft paper's RPC descriptions.

RemoteProcedureResponse
Data interchange format for RPC responses. * Accepted mean that it worked. * Rejected means that rpc.term < node.persistent_state.current_term or if the Node's log doesn't contain the entry at rpc.prev_log_index that maches prev_log_term. The caller should follow the current_leader it is directed to. The UUID should match the coresponding RPC.