Enum raft::interchange::ClientRequest

pub enum ClientRequest<T> {
    IndexRange(IndexRange),
    AppendRequest(AppendRequest<T>),
}

Data interchange request format for Client <-> Node Communication. Each variant of this is a command which can be asked of the RaftNode after it is spun up with node.spinup() The node attached to this application will poll it's channel regularly and return results on the channel. If you're wondering where vote requesting is, it's hidden within the module. TODO: Currently requests are not queued or gaurenteed to serviced in order. This is probably a bad thing as most clients will .send() then .recv(). We can probably make a service queue for this.

Variants

IndexRange

Gets the log entries from start to end.

AppendRequest

Asks the node to append an entry after a given entry.

Methods

impl<T> ClientRequest<T>

fn index_range(start: u64, end: u64) -> ClientRequest<T>

Returns (term, success)

fn append_request(prev_log_index: u64, prev_log_term: u64, entries: Vec<T>) -> ClientRequest<T>

Returns (term, voteGranted)