Struct raft::kvraft::client::Clerk

pub struct Clerk {
    pub name: String,
    pub servers: Vec<KvClient>,
    id: String,
    next_seq: AtomicU64,
    last_leader: AtomicUsize,
}

A key-value clerk.

Fields
name: Stringservers: Vec<KvClient>

All accessible kv servers.

id: Stringnext_seq: AtomicU64

Sequence number for next request.

last_leader: AtomicUsize

Index of the last remembered leader in servers.

Implementations

impl Clerk

pub fn new(name: String, servers: Vec<KvClient>) -> Clerk

fn next_seq(&self) -> u64

Get the next sequence number and increase it.

fn cycle_servers(&self) -> impl Iterator<Item = (usize, &KvClient)>

An infinite iterator of all servers, start from last_leader.

async fn call(&self, args: KvRequest) -> String

Call servers for given args repeatedly until success.

pub async fn get_async(&self, key: String) -> String

Async version of get request.

Note that we cannot call block_on if we already entered a executor. Thus, this method should be called in any async context.

pub fn get(&self, key: String) -> String

fetch the current value for a key. returns “” if the key does not exist. keeps trying forever in the face of all other errors.

pub async fn put_async(&self, key: String, value: String)

Async version of put request.

Note that we cannot call block_on if we already entered a executor. Thus, this method should be called in any async context.

pub fn put(&self, key: String, value: String)

Put a key-value pair into kv servers.

pub async fn append_async(&self, key: String, value: String)

Async version of append request.

Note that we cannot call block_on if we already entered a executor. Thus, this method should be called in any async context.

pub fn append(&self, key: String, value: String)

Append the value to the entry for key. Will do insertion if the entry not exists.