Module raft::log

Types related to Raft log storage.

Raft requires a backing storage for entries of its distributed log as they are being replicated to and from other nodes. The RaftLog trait is implemented for that purpose, and the implementation is supplied to RaftNode.

Modules

mem
A naive in-memory implementation of RaftLog, primarily for testing.

Structs

CommittedIter
An iterator yielding committed log entries.

Traits

RaftLog
An interface for storage of the Raft log of a RaftNode.