Trait raft::log::Log

pub trait Log {
    type Command: Clone;
    fn append(&mut self, term: Term, command: Self::Command) -> Index;

    fn truncate_if_different_and_append(
        &mut self, 
        prev_index: Index, 
        items: impl IntoIterator<Item = Item<Self::Command>>
    ) -> Index;

    fn log_term_matches(&self, index: Index, term: Term) -> bool;

    fn last_log_term(&self) -> Term;

    fn last_log_index(&self) -> Index;

    fn get_command(&self, idx: Index) -> &Self::Command;

    fn get_term(&self, idx: Index) -> Option<Term>;

    fn get_from(&self, idx: Index) -> Vec<Item<Self::Command>>;

    fn cmp(
        &self, 
        other_last_log_index: Index, 
        other_last_log_term: Term
    ) -> Ordering { ... }
}

Associated Types

type Command: Clone

Loading content...
Required methods

fn append(&mut self, term: Term, command: Self::Command) -> Index

Add a new command and return its index

fn truncate_if_different_and_append(
&mut self,
prev_index: Index,
items: impl IntoIterator<Item = Item<Self::Command>>
) -> Index

Check items against existing items in log, remove from log any conflicting items and those following them and append the remaining new items.

Returns

Index of the last new item added #Panics If prev_index doesn't point at an existing log entry

fn log_term_matches(&self, index: Index, term: Term) -> bool

Check whether the log entry's term at the specified index matches

fn last_log_term(&self) -> Term

Term of the last log item

fn last_log_index(&self) -> Index

Index of the last log item

fn get_command(&self, idx: Index) -> &Self::Command

Command at index

fn get_term(&self, idx: Index) -> Option<Term>

Term at index

fn get_from(&self, idx: Index) -> Vec<Item<Self::Command>>

All items following and including the index

Loading content...
Provided methods

fn cmp(
&self,
other_last_log_index: Index,
other_last_log_term: Term
) -> Ordering

Raft determines which of two logs is more up-to-date by comparing the index and term of the last entries in the logs. If the logs have last entries with different terms, then the log with the later term is more up-to-date. If the logs end with the same term, then whichever log is longer is more up-to-date.

Loading content...