Module raft::persistent_log

The persistent storage of Raft state.

In your consuming application you will want to implement this trait on one of your structures. This could adapt to a database, a file, or even just POD.

Note: Your consuming application should not necessarily interface with this data. It is meant for internal use by the library, we simply chose not to be opinionated about how data is stored.

Structs

FsLog
Stores log on disk as 8 bytes for the version identifier, 8 bytes for current_term, 8 bytes for voted_for, and as much as needed for the log. Each log entry is stored as an 8 byte length specifier which is the total length of the entry in bytes, including the length specifier, followed by 8 bytes specifying the term, plus a variable length entry, which is the serialized command sent to raft by the client.

MemLog
This is a Log implementation that stores entries in a simple in-memory vector. Other data is stored in a struct. It is chiefly intended for testing.

Enums

Error
Non-instantiable error type for MemLog

Traits

Log
A store of persistent Raft state.