Struct raft::persistent_log::FsLog

pub struct FsLog { /* fields omitted */ }

Stores log on disk as 8 bytes for the version identifier, 8 bytes for current_term, 8 bytes for voted_for, and as much as needed for the log. Each log entry is stored as an 8 byte length specifier which is the total length of the entry in bytes, including the length specifier, followed by 8 bytes specifying the term, plus a variable length entry, which is the serialized command sent to raft by the client.

Methods

impl FsLog

fn new(filename: &Path) -> Result<FsLog, Error>