Trait raft::persistent_log::Log

pub trait Log: Clone + Debug + Send + 'static {
    type Error: Error + Debug + Sized + 'static;
    fn current_term(&self) -> Result<Term, Self::Error>;
    fn set_current_term(&mut self, term: Term) -> Result<(), Self::Error>;
    fn inc_current_term(&mut self) -> Result<Term, Self::Error>;
    fn voted_for(&self) -> Result<Option<ServerId>, Self::Error>;
    fn set_voted_for(&mut self, server: ServerId) -> Result<(), Self::Error>;
    fn latest_log_index(&self) -> Result<LogIndex, Self::Error>;
    fn latest_log_term(&self) -> Result<Term, Self::Error>;
    fn entry(&self, index: LogIndex) -> Result<(Term, &[u8]), Self::Error>;
    fn append_entries(
        &mut self, 
        from: LogIndex, 
        entries: &[(Term, &[u8])]
    ) -> Result<(), Self::Error>;

    fn entries(
        &self, 
        lo: LogIndex, 
        hi: LogIndex
    ) -> Result<Vec<(Term, &[u8])>, Self::Error> { ... }
}

A store of persistent Raft state.

Associated Types

type Error: Error + Debug + Sized + 'static

Required Methods

fn current_term(&self) -> Result<Term, Self::Error>

Returns the latest known term.

fn set_current_term(&mut self, term: Term) -> Result<(), Self::Error>

Sets the current term to the provided value. The provided term must be greater than the current term. The voted_for value will be reset`.

fn inc_current_term(&mut self) -> Result<Term, Self::Error>

Increment the current term. The voted_for value will be reset.

fn voted_for(&self) -> Result<Option<ServerId>, Self::Error>

Returns the candidate id of the candidate voted for in the current term (or none).

fn set_voted_for(&mut self, server: ServerId) -> Result<(), Self::Error>

Sets the candidate id voted for in the current term.

fn latest_log_index(&self) -> Result<LogIndex, Self::Error>

Returns the index of the latest persisted log entry (0 if the log is empty).

fn latest_log_term(&self) -> Result<Term, Self::Error>

Returns the term of the latest persisted log entry (0 if the log is empty).

fn entry(&self, index: LogIndex) -> Result<(Term, &[u8]), Self::Error>

Returns the entry at the provided log index.

fn append_entries(
&mut self,
from: LogIndex,
entries: &[(Term, &[u8])]
) -> Result<(), Self::Error>

Appends the provided entries to the log beginning at the given index.

Provided Methods

fn entries(
&self,
lo: LogIndex,
hi: LogIndex
) -> Result<Vec<(Term, &[u8])>, Self::Error>

Returns the given range of entries (excluding the right endpoint).