Module raft::raft

Modules

errors

persister

Support for Raft and kvraft to save persistent Raft state (log &c) and k/v server snapshots.

states

Macros

rlog

Macro for logging message combined with state of the Raft peer.

Structs

Node

Raft service which triggers event loop, timer and RPC handlers for inner Raft instance.

Raft

A single Raft peer.

State

State of a raft peer (for testing).

Enums

ApplyMsg

Message sent to the service.

Event

Event run by a Raft peer in a event loop.

TimerAction

Action for Raft peer to control the timer.

Constants

HEART_BEAT_MILLIS

Raft peer heart beat interval in millis, should not be too small to ensure passing tests.

Type Definitions

RpcResult