Struct raft::raft::Node

pub struct Node {
    raft: Arc<RwLock<Raft>>,
    event_loop_tx: UnboundedSender<Event>,
    shutdown_tx: Arc<Mutex<Option<Sender<()>>>>,
    pub executor: Executor,
}

Raft service which triggers event loop, timer and RPC handlers for inner Raft instance.

Fields
raft: Arc<RwLock<Raft>>

Inner Raft instance.

event_loop_tx: UnboundedSender<Event>

Commit event to handle asynchronously, used by timer

shutdown_tx: Arc<Mutex<Option<Sender<()>>>>

Notify the event loop to shutdown

executor: Executor

Thread pool executor shared with inner Raft instance.

Implementations

impl Node

pub fn new(raft: Raft) -> Node

Create a new raft service.

fn start_event_loop(
&self,
event_loop_rx: UnboundedReceiver<Event>,
shutdown_rx: Receiver<()>
)

Main event loop for the Raft state machine. Receive events from event_loop_rx indefinitely, and call the event handler with MUTABLE reference.

Receiving message from shutdown_rx will break the loop and then cause event_loop_rx to be dropped, which finally make all other tasks holding the event_loop_tx to gracefully exit.

fn start_timer(&self, timer_action_rx: UnboundedReceiver<TimerAction>)

Timer task for the Raft state machine: election timeout and heart beat. Can be controlled by actions from timer_action_rx.

All timer events will be sent through event_loop_tx and then processed by the event loop. Thus the timer expects the event loop not to be so “crowded” to ensure the real-time.

pub fn start<M>(&self, command: &M) -> Result<(u64, u64)> where
M: Message,

the service using Raft (e.g. a k/v server) wants to start agreement on the next command to be appended to Raft’s log. if this server isn’t the leader, returns Error::NotLeader. otherwise start the agreement and return immediately. there is no guarantee that this command will ever be committed to the Raft log, since the leader may fail or lose an election. even if the Raft instance has been killed, this function should return gracefully.

the first value of the tuple is the index that the command will appear at if it’s ever committed. the second is the current term.

This method must return without blocking on the raft.

pub fn term(&self) -> u64

The current term of this peer.

pub fn is_leader(&self) -> bool

Whether this peer believes it is the leader.

pub fn get_state(&self) -> State

The current state of this peer.

pub fn snapshot(&mut self, index: u64, snapshot: Vec<u8>)

the service says it has created a snapshot that has all info up to and including index. this means the service no longer needs the log through (and including) that index. Raft should now trim its log as much as possible.

pub fn cond_install_snapshot(
&mut self,
last_included_term: u64,
last_included_index: u64,
snapshot: Vec<u8>
) -> bool

A service wants to switch to snapshot. Only do so if Raft hasn’t have more recent info since it communicate the snapshot on applyCh.

pub fn kill(&self)

the tester calls kill() when a Raft instance won’t be needed again. you are not required to do anything in kill(), but it might be convenient to (for example) turn off debug output from this instance. In Raft paper, a server crash is a PHYSICAL crash, A.K.A all resources are reset. But we are simulating a VIRTUAL crash in tester, so take care of background threads you generated with this Raft Node.