Struct raft::raft::Raft

pub struct Raft {
    peers: Vec<RaftClient>,
    me: usize,
    role: RoleState,
    p: PersistentStateWrapper,
    v: VolatileState,
    event_loop_tx: Option<UnboundedSender<Event>>,
    timer_action_tx: Option<UnboundedSender<TimerAction>>,
    apply_tx: UnboundedSender<ApplyMsg>,
    executor: Executor,
}

A single Raft peer.

Fields
peers: Vec<RaftClient>

RPC end points of all peers.

me: usize

This peer’s index into peers[].

role: RoleState

Role of this peer with its specific states.

p: PersistentStateWrapper

Persistent state: current_form, voted_for, log.

v: VolatileState

Volatile state: commit_index, last_applied.

event_loop_tx: Option<UnboundedSender<Event>>

Commit event to handle asynchronously, should always be Some.

timer_action_tx: Option<UnboundedSender<TimerAction>>

Commit timer action to handle asynchronously, like to reset election timeout, should always be Some.

apply_tx: UnboundedSender<ApplyMsg>

Apply commands to the application.

executor: Executor

Shared thread pool executor for background RPC requests.

Implementations

impl Raft

pub fn new(
peers: Vec<RaftClient>,
me: usize,
persister: Box<dyn Persister>,
apply_tx: UnboundedSender<ApplyMsg>
) -> Raft

the service or tester wants to create a Raft server. the ports of all the Raft servers (including this one) are in peers. this server’s port is peers[me]. all the servers’ peers arrays have the same order. persister is a place for this server to save its persistent state, and also initially holds the most recent saved state, if any. apply_ch is a channel on which the tester or service expects Raft to send ApplyMsg messages. This method must return quickly.

fn persist(&self, snapshot: Option<Vec<u8>>)

save Raft’s persistent state to stable storage, where it can later be retrieved after a crash and restart. see paper’s Figure 2 for a description of what should be persistent.

fn start<M>(&mut self, command: &M) -> Result<(u64, u64)> where
M: Message,

The service using Raft (e.g. a k/v server) wants to start agreement on the next command to be appended to Raft’s log. Returns immediately with (index, term) of the command entry.

impl Raft

fn event_loop_tx(&self) -> &UnboundedSender<Event>

Get a ref to event_loop_tx.

fn timer_action_tx(&self) -> &UnboundedSender<TimerAction>

Get a ref to timer_action_tx.

fn other_peers(&self) -> impl Iterator<Item = (usize, &RaftClient)>

Enumerated peers except me.

fn append_entries_args(&self, start_at: usize) -> Option<AppendEntriesArgs>

Make arguments for AppendEntries RPC by the index where entries start_at.

fn request_vote_args(&self) -> RequestVoteArgs

Make arguments for RequestVote RPC.

fn install_snapshot_args(&self) -> InstallSnapshotArgs

Make arguments for InstallSnapshot RPC.

impl Raft

fn update_term(&mut self, term: u64)

Update current term and clear voted_for state. WILL NOT touch the role state.

fn turn(&mut self, role: RoleState)

Turn to new role.

fn turn_follower(&mut self)

Turn to follower.

fn turn_candidate(&mut self)

Turn to candidate, with 1 votes from ourself.

fn turn_leader(&mut self)

Turn to leader, with reinitialized leader states.

impl Raft

fn start_new_election(&self)

Start a new election by requesting votes from all other peers. Will return immediately, while push all replies into event loop asynchronously.

fn heart_beat_sync_log(&self)

Sync log by appending entries OR installing snapshot to all other peers. Will return immediately, while push all replies into event loop asynchronously.

Should ONLY be called by leader on heart beat.

fn commit_and_apply_up_to(&mut self, new_commit_index: usize)

Update commit_index and commit (and apply) all log entries up to it.

fn leader_try_commit(&mut self)

Try to traverse leader’s start_indexs to find new entries that are replicated and commit them.

Should ONLY be called by leader on receiving AppendEntries replies.

fn reset_timeout(&self)

Tell timer to reset the election timeout.

fn cond_install_peer_snapshot(
&mut self,
last_included_term: u64,
last_included_index: usize,
snapshot: Vec<u8>
) -> bool

Service want Raft peer to conditionally install snapshot given from peer and applied by us previously.

Returns whether we grant to install this snapshot.

fn install_snapshot(
&mut self,
from_peer: bool,
included_index: usize,
included_term: u64,
snapshot: Vec<u8>
) -> bool

Compact the log and set commit_index up to the last index of the snapshot.

Will do nothing if the compaction is failed. Returns whether the compaction is successful.

impl Raft

fn handle_event(&mut self, event: Event)

Entrypoint for handling event received from the event loop.

fn handle_election_timeout(&mut self)

Election timeout. Will turn to candidate and start new election.

fn handle_heart_beat(&self)

Heart beat. Sync log if we are leader.

fn handle_request_vote_request(
&mut self,
args: RequestVoteArgs
) -> Result<RequestVoteReply>

RPC request for RequestVote.

fn handle_request_vote_reply(
&mut self,
from: usize,
reply: Result<RequestVoteReply>
)

Reply from our prior RequestVote RPC request.

fn handle_append_entries_request(
&mut self,
args: AppendEntriesArgs
) -> Result<AppendEntriesReply>

RPC request for AppendEntries.

fn handle_append_entries_reply(
&mut self,
from: usize,
reply: Result<AppendEntriesReply>,
new_next_index: usize
)

Reply from our prior AppendEntries RPC request.

fn handle_install_snapshot_request(
&mut self,
args: InstallSnapshotArgs
) -> Result<InstallSnapshotReply>

RPC request for InstallSnapshot.

fn handle_install_snapshot_reply(
&mut self,
from: usize,
reply: Result<InstallSnapshotReply>,
new_next_index: usize
)

Reply from our prior InstallSnapshot RPC request.