Struct raft::server::Server

pub struct Server<Command: Clone, Log: Log<Command = Command>> { /* fields omitted */ }

The logic of a raft server except for delay, timouts, messaging and configuration which are handled by the client

Implementations

impl<Command: 'static + Clone, Log: Log<Command = Command> + Default> Server<Command, Log>

pub fn new(
server_id: ServerId,
persistent_state: Persistent<Log>,
state_machine: Box<Receiver<Command>>
) -> Self

impl<Command: 'static + Clone, Log: Log<Command = Command>> Server<Command, Log>

pub fn election_timeout(&mut self) -> RequestVoteRequest

No messages have been received over the election timeout.

Followers (5.2):

-- If election timeout elapses without receiving AppendEntriesRPC from current leader or granting vote to candidate: convert to candidate

Candidates (5.2):

On conversion to candidate, start election: -- Increment currentTerm -- Vote for self -- Reset election timer -- Send RequestVote RPCs to all other servers ... -- If election timeout elapses: start new election

pub fn receive_append_entries<LogEntries: IntoIterator<Item = Item<Log::Command>>>(
&mut self,
req: AppendEntriesRequest<Log::Command, LogEntries>
) -> AppendEntriesResponse

While waiting for votes, a candidate may receive an AppendEntries RPC from another server claiming to be leader. If the leader’s term (included in its RPC) is at least as large as the candidate’s current term, then the candidate recognizes the leader as legitimate and returns to follower state. If the term in the RPC is smaller than the candidate’s current term, then the candidate rejects the RPC and continues in candidate state.

pub fn receive_request_vote(
&mut self,
req: RequestVoteRequest
) -> RequestVoteResponse

pub fn receive_vote(
&mut self,
res: RequestVoteResponse,
total_servers: ServerId
) -> Option<AppendEntriesRequest<Log::Command, Empty<Item<Log::Command>>>>

Receives a RequestVote response

Return

if the server determines it has won the election with the received vote, an AppendEntriesRequest heartbeat is returned which the client needs to send to all other servers in the cluster. At this point the client ust also start a heartbeat timer and start calling the heartbeat method on it

pub fn receive_append_entries_response(
&mut self,
from: ServerId,
match_index: Index,
res: AppendEntriesResponse
) -> Vec<CommandPtr>

Receive response from AppendEntriesRequest

If successful: update nextIndex and matchIndex for follower (5.3)

If AppendEntriesRequest fails because of log inconsistency:decrement nextIndex and retry (5.3)

Args

Requires the id of the server responding the the append entries request, and the match_index which is the index of the last log entry sent in the request.

Return

Vec of command pointers which have been newly applied, the client will need to send any client responses associated with those commands

pub fn command(
&mut self,
command: Log::Command
) -> Result<CommandPtr, NotLeader>

Client adds a command

Return

Command pointer, the client can be informed of the successful application of the command once this command pointer s returned by receive_append_entries_response

pub fn heartbeat(
&mut self
) -> Vec<(ServerId, AppendEntriesRequest<Command, Vec<Item<Command>>>)>

Call this on a timer from when an election is won, until a pre_handle indicates that a new election has been called

impl<Command: 'static + Clone, Log: Log<Command = Command>> Server<Command, Log>

pub fn get_state(&self) -> &ServerState<Log>