Module raft::state

Structs

LeaderState
Leader Only Reinitialized after election.

PersistentState
Persistent state Must be updated to stable storage before RPC response.

Transaction

VolatileState
Volatile state

Enums

NodeState
Nodes can either be:

TransactionState
Used to signify the state of of a Request/Response pair. This is only needed on the original sender... not on the reciever.