Enum raft::state::NodeState

pub enum NodeState {
    Follower(VecDeque<Transaction>),
    Leader(LeaderState),
    Candidate(Vec<Transaction>),
}

Nodes can either be:

A Follower, which replicates AppendEntries requests and votes for it's leader.

A Leader, which leads the cluster by serving incoming requests, ensuring data is replicated, and issuing heartbeats..

A Candidate, which campaigns in an election and may become a Leader (if it gets enough votes) or a Follower, if it hears from a Leader.

Variants

Follower

Leader

Candidate