Struct raft::state::PersistentState

pub struct PersistentState<T: Encodable + Decodable + Send + Clone> {
    // some fields omitted
}

Persistent state Must be updated to stable storage before RPC response.

Methods

impl<T: Encodable + Decodable + Send + Clone> PersistentState<T>

fn new(current_term: u64, log_path: Path) -> PersistentState<T>

fn get_last_index(&self) -> u64

Gets the last_index which you can use to make append requests with.

fn get_last_term(&self) -> u64

fn get_current_term(&self) -> u64

Gets the current_term which is used for request vote.

fn set_current_term(&mut self, new: u64) -> Result<()>

Sets the current_term. This should reflect on stable storage.

fn inc_current_term(&mut self)

Increments the current_term. This should reflect on stable storage.

fn get_voted_for(&mut self) -> Option<u64>

Gets the voted_for.

fn set_voted_for(&mut self, new: Option<u64>) -> Result<()>

Sets the `voted_for. This should reflect on stable storage.

fn append_entries(&mut self, prev_log_index: u64, prev_log_term: u64, entries: Vec<(u64, T)>) -> Result<()>

fn purge_from_index(&mut self, from_line: u64) -> Result<()>

Removes all entries from from to the last entry, inclusively.

fn retrieve_entries(&mut self, start: u64, end: u64) -> Result<Vec<(u64, T)>>

fn retrieve_entry(&mut self, index: u64) -> Result<(u64, T)>