Struct raft::state::ServerState

pub struct ServerState<Log> { /* fields omitted */ }

Volatile state on all servers

Implementations

impl<Log: Log> ServerState<Log>

pub fn new(persistent_state: Persistent<Log>) -> Self

pub fn current_term(&self) -> Term

pub fn voted_for(&self) -> Option<ServerId>

pub fn commit_index(&self) -> Index

pub fn is_follower(&self) -> bool

pub fn is_candidate(&self) -> bool

pub fn is_leader(&self) -> bool

pub fn follow_new_term(&mut self, term: Term)

Another server is acknowledged as leader

pub fn start_election(&mut self, server_id: ServerId) -> RequestVoteRequest

No messages have been received over the election timeout. Start a new election term

Followers (5.2):

-- If election timeout elapses without receiving AppendEntriesRPC from current leader or granting vote to candidate: convert to candidate

Candidates (5.2):

On conversion to candidate, start election: -- Increment currentTerm -- Vote for self -- Reset election timer (to be handled by caller) -- Send RequestVote RPCs to all other servers

pub fn become_leader(&mut self, own_id: ServerId, num_servers: ServerId)

Win an election and become a leader

pub fn set_commit_index(&mut self, commit_index: Index)

Panics

In case the term tries to decrease

pub fn last_applied(&self) -> Index

pub fn add_command(
&mut self,
command: Log::Command
) -> Result<CommandPtr, NotLeader>

Add a command from a client to this leader

pub fn set_last_applied(&mut self, last_applied: Index)

Panics

In case the term tries to decrease

pub fn produce_append_entries(
&mut self,
leader_id: ServerId
) -> Vec<(ServerId, AppendEntriesRequest<Log::Command, Vec<Item<Log::Command>>>)>

Prepare heartbeat / log updates for followers #Panics if not a leader

pub fn update_follower(&mut self, from: ServerId, match_idx: Index)

A follower has successfully appended If successful: update nextIndex and matchIndex for follower (5.3)

pub fn follower_inconsistent(&mut self, from: ServerId)

A followers response indicates that they are inconsistent with the leader If AppendEntries fails because of log inconsistency: decrement nextIndex and retry (5.3)

pub fn receive_append_entries<LogEntries: IntoIterator<Item = Item<Log::Command>>>(
&mut self,
req: AppendEntriesRequest<Log::Command, LogEntries>
) -> AppendEntriesResponse

Invoked by leader to replicate log entries (5.3); also used as heartbeat (5.2).

Reply false if term < currentTerm (5.1)

Reply false if log doesn’t contain an entry at prevLogIndex whose term matches prevLogTerm (5.3)

If an existing entry conflicts with a new one (same index but different terms), delete the existing entry and all that follow it (5.3)

Append any new entries not already in the log

If leaderCommit > commitIndex, set commitIndex = min(leaderCommit, index of last new entry)

pub fn receive_request_vote(
&mut self,
req: RequestVoteRequest
) -> RequestVoteResponse

Requested to vote for candidate

Reply false if term < currentTerm (5.1)

If votedFor is null or candidateId, and candidate’s log is at least as up-to-date as receiver’s log, grant vote (5.2, 5.4)

pub fn apply_commited(
&mut self,
receiver: &mut Receiver<Log::Command>
) -> Vec<(Index, Term)>

Apply all commited log items, return command pointers to those commands newly applied

impl<Log: Log> ServerState<Log>

pub fn log(&self) -> &Log

pub fn get_leader_state(&self) -> &Leader