Module raft::state_machine

The state machine which Raft applied commited entries too.

This trait is meant to be implemented such that the commands issued to it via apply() will be reflected in your consuming application. Commands send via apply() have been committed in the cluser. Unlike store, your application should consume data produced by this and accept it as truth.

Structs

ChannelStateMachine
A state machine that simply redirects all commands to a channel.

NullStateMachine
A state machine with no states.

Traits

StateMachine