Struct raft::storage::MemStorage

pub struct MemStorage { /* fields omitted */ }

MemStorage is a thread-safe but incomplete implementation of Storage, mainly for tests.

A real Storage should save both raft logs and applied data. However MemStorage only contains raft logs. So you can call MemStorage::append to persist new received unstable raft logs and then access them with Storage APIs. The only exception is Storage::snapshot. There is no data in Snapshot returned by MemStorage::snapshot because applied data is not stored in MemStorage.

Implementations

impl MemStorage

pub fn new() -> MemStorage

Returns a new memory storage value.

pub fn new_with_conf_state<T>(conf_state: T) -> MemStorage where
ConfState: From<T>,

Create a new MemStorage with a given Config. The given Config will be used to initialize the storage.

You should use the same input to initialize all nodes.

pub fn initialize_with_conf_state<T>(&self, conf_state: T) where
ConfState: From<T>,

Initialize a MemStorage with a given Config.

You should use the same input to initialize all nodes.

pub fn rl(&self) -> RwLockReadGuard<'_, MemStorageCore>

Opens up a read lock on the storage and returns a guard handle. Use this with functions that don’t require mutation.

pub fn wl(&self) -> RwLockWriteGuard<'_, MemStorageCore>

Opens up a write lock on the storage and returns guard handle. Use this with functions that take a mutable reference to self.