Struct raft::storage::MemStorageCore

pub struct MemStorageCore { /* fields omitted */ }

Methods

impl MemStorageCore

pub fn set_hardstate(&mut self, hs: HardState)

set_hardstate saves the current HardState.

pub fn apply_snapshot(&mut self, snapshot: Snapshot) -> Result<()>

apply_snapshot overwrites the contents of this Storage object with those of the given snapshot.

pub fn create_snapshot(
&mut self,
idx: u64,
cs: Option<ConfState>,
data: Vec<u8>
) -> Result<&Snapshot>

create_snapshot makes a snapshot which can be retrieved with snapshot() and can be used to reconstruct the state at that point. If any configuration changes have been made since the last compaction, the result of the last apply_conf_change must be passed in.

pub fn compact(&mut self, compact_index: u64) -> Result<()>

compact discards all log entries prior to compact_index. It is the application's responsibility to not attempt to compact an index greater than RaftLog.applied.

pub fn append(&mut self, ents: &[Entry]) -> Result<()>

Append the new entries to storage. TODO: ensure the entries are continuous and entries[0].get_index() > self.entries[0].get_index()