Struct raft::Client

pub struct Client {
    pub id: ClientId,
    // some fields omitted
}

The representation of a Client connection to the cluster.

Fields
id: ClientId

The Uuid of the client, should be unique in the cluster.

Methods

impl Client

fn new(cluster: HashSet<SocketAddr>) -> Client

Creates a new client.

fn propose(&mut self, entry: &[u8]) -> Result<Vec<u8>>

Proposes an entry to be appended to the replicated log. This will only return once the entry has been durably committed. Returns Error when the entire cluster has an unknown leader. Try proposing again later.

fn query(&mut self, query: &[u8]) -> Result<Vec<u8>>

Queries an entry from the state machine. This is non-mutating and doesn't go through the durable log. Like .propose() this will only communicate with the leader of the cluster.