Struct raft::Config

pub struct Config {
    pub id: u64,
    pub peers: Vec<u64>,
    pub learners: Vec<u64>,
    pub election_tick: usize,
    pub heartbeat_tick: usize,
    pub applied: u64,
    pub max_size_per_msg: u64,
    pub max_inflight_msgs: usize,
    pub check_quorum: bool,
    pub pre_vote: bool,
    pub read_only_option: ReadOnlyOption,
    pub skip_bcast_commit: bool,
    pub tag: String,
}

Config contains the parameters to start a raft.

Fields
id: u64

id is the identity of the local raft. ID cannot be 0.

peers: Vec<u64>

peers contains the IDs of all nodes (including self) in the raft cluster. It should only be set when starting a new raft cluster. Restarting raft from previous configuration will panic if peers is set. peer is private and only used for testing right now.

learners: Vec<u64>

learners contains the IDs of all learner nodes (maybe include self if the local node is a learner) in the raft cluster. learners only receives entries from the leader node. It does not vote or promote itself.

election_tick: usize

ElectionTick is the number of node.tick invocations that must pass between elections. That is, if a follower does not receive any message from the leader of current term before ElectionTick has elapsed, it will become candidate and start an election. election_tick must be greater than HeartbeatTick. We suggest election_tick = 10 * HeartbeatTick to avoid unnecessary leader switching

heartbeat_tick: usize

HeartbeatTick is the number of node.tick invocations that must pass between heartbeats. That is, a leader sends heartbeat messages to maintain its leadership every heartbeat ticks.

applied: u64

Applied is the last applied index. It should only be set when restarting raft. raft will not return entries to the application smaller or equal to Applied. If Applied is unset when restarting, raft might return previous applied entries. This is a very application dependent configuration.

max_size_per_msg: u64

MaxSizePerMsg limits the max size of each append message. Smaller value lowers the raft recovery cost(initial probing and message lost during normal operation). On the other side, it might affect the throughput during normal replication. Note: math.MaxUusize64 for unlimited, 0 for at most one entry per message.

max_inflight_msgs: usize

max_inflight_msgs limits the max number of in-flight append messages during optimistic replication phase. The application transportation layer usually has its own sending buffer over TCP/UDP. Setting MaxInflightMsgs to avoid overflowing that sending buffer. TODO: feedback to application to limit the proposal rate?

check_quorum: bool

check_quorum specifies if the leader should check quorum activity. Leader steps down when quorum is not active for an electionTimeout.

pre_vote: bool

pre_vote enables the Pre-Vote algorithm described in raft thesis section 9.6. This prevents disruption when a node that has been partitioned away rejoins the cluster.

read_only_option: ReadOnlyOption

read_only_option specifies how the read only request is processed.

skip_bcast_commit: bool tag: String

tag is only used for logging

Methods

impl Config

pub fn validate(&self) -> Result<()>