Struct raft::ProgressSet

pub struct ProgressSet { /* fields omitted */ }

ProgressSet contains several Progresses, which could be Leader, Follower and Learner.

Implementations

impl ProgressSet

pub fn configuration(&self) -> &Configuration

The current configuration state of the cluster.

impl ProgressSet

pub fn new(logger: Logger) -> Self

Creates a new ProgressSet.

pub fn with_capacity(voters: usize, learners: usize, logger: Logger) -> Self

Create a progress set with the specified sizes already reserved.

pub fn enable_group_commit(&mut self, enable: bool)

Configures group commit.

pub fn group_commit(&self) -> bool

Whether enable group commit.

pub fn voters(&self) -> impl Iterator<Item = (&u64, &Progress)>

Returns the status of voters.

Note: Do not use this for majority/quorum calculation. The Raft node may be transitioning to a new configuration and have two qourums. Use has_quorum instead.

pub fn learners(&self) -> impl Iterator<Item = (&u64, &Progress)>

Returns the status of learners.

Note: Do not use this for majority/quorum calculation. The Raft node may be transitioning to a new configuration and have two qourums. Use has_quorum instead.

pub fn voters_mut(&mut self) -> impl Iterator<Item = (&u64, &mut Progress)>

Returns the mutable status of voters.

Note: Do not use this for majority/quorum calculation. The Raft node may be transitioning to a new configuration and have two qourums. Use has_quorum instead.

pub fn learners_mut(&mut self) -> impl Iterator<Item = (&u64, &mut Progress)>

Returns the mutable status of learners.

Note: Do not use this for majority/quorum calculation. The Raft node may be transitioning to a new configuration and have two qourums. Use has_quorum instead.

pub fn voter_ids(&self) -> HashSet<u64, BuildHasherDefault<FxHasher>>

Returns the ids of all known voters.

Note: Do not use this for majority/quorum calculation. The Raft node may be transitioning to a new configuration and have two qourums. Use has_quorum instead.

pub fn learner_ids(&self) -> HashSet<u64, BuildHasherDefault<FxHasher>>

Returns the ids of all known learners.

Note: Do not use this for majority/quorum calculation. The Raft node may be transitioning to a new configuration and have two qourums. Use has_quorum instead.

pub fn get(&self, id: u64) -> Option<&Progress>

Grabs a reference to the progress of a node.

pub fn get_mut(&mut self, id: u64) -> Option<&mut Progress>

Grabs a mutable reference to the progress of a node.

pub fn iter(&self) -> impl ExactSizeIterator<Item = (&u64, &Progress)>

Returns an iterator across all the nodes and their progress.

Note: Do not use this for majority/quorum calculation. The Raft node may be transitioning to a new configuration and have two qourums. Use has_quorum instead.

pub fn iter_mut(
&mut self
) -> impl ExactSizeIterator<Item = (&u64, &mut Progress)>

Returns a mutable iterator across all the nodes and their progress.

Note: Do not use this for majority/quorum calculation. The Raft node may be transitioning to a new configuration and have two qourums. Use has_quorum instead.

pub fn insert_voter(&mut self, id: u64, pr: Progress) -> Result<()>

Adds a voter to the group.

Errors

id is in the voter set.

id is in the learner set.

pub fn insert_learner(&mut self, id: u64, pr: Progress) -> Result<()>

Adds a learner to the group.

Errors

id is in the voter set.

id is in the learner set.

pub fn remove(&mut self, id: u64) -> Result<Option<Progress>>

Removes the peer from the set of voters or learners.

Errors

pub fn promote_learner(&mut self, id: u64) -> Result<()>

Promote a learner to a peer.

pub fn maximal_committed_index(&mut self) -> (u64, bool)

Returns the maximal committed index for the cluster. The bool flag indicates whether the index is computed by group commit algorithm successfully.

Eg. If the matched indexes are [2,2,2,4,5], it will return 2. If the matched indexes and groups are [(1, 1), (2, 2), (3, 2)], it will return 1.

pub fn candidacy_status<'a>(
&self,
votes: impl IntoIterator<Item = (&'a u64, &'a bool)>
) -> CandidacyStatus

Returns the Candidate's eligibility in the current election.

If it is still eligible, it should continue polling nodes and checking. Eventually, the election will result in this returning either Elected or Ineligible, meaning the election can be concluded.

pub fn quorum_recently_active(&mut self, perspective_of: u64) -> bool

Determines if the current quorum is active according to the this raft node. Doing this will set the recent_active of each peer to false.

This should only be called by the leader.

pub fn has_quorum(
&self,
potential_quorum: &HashSet<u64, BuildHasherDefault<FxHasher>>
) -> bool

Determine if a quorum is formed from the given set of nodes.

This is the only correct way to verify you have reached a quorum for the whole group.