Struct raft::Raft

pub struct Raft {
    // some fields omitted
}

This is the primary interface with a Server in the cluster.

Note: Creating a new Raft client will, for now, automatically spawn a Server with the relevant parameters. This may be changed in the future. This is based on the assumption that any consuming application interacting with a Raft cluster will also be a participant.

Methods

impl Raft

fn new<S, M>(addr: SocketAddr, cluster_members: HashSet<SocketAddr>, store: S, state_machine: M) -> Raft where S: Store, M: StateMachine

Create a new Raft client that has a cooresponding Server attached. Note that this Raft may not necessarily interact with the cooreponding Server, it will interact with the Leader of a cluster in almost all cases. Note: All requests are blocking, by design from the Raft paper.

fn append(&mut self, entry: &[u8]) -> Result<()>

Appends an entry to the replicated log. This will only return once it's properly replicated to a majority of nodes.

fn die(&mut self, target: SocketAddr, reason: String) -> Result<()>

Kills the node. Should only really be used for testing purposes. Accepts a SocketAddr because if you're going to kill a node you should be able to pick your victim.

fn refresh_leader(&mut self) -> Result<()>

This function will force the Raft interface to refresh it's knowledge of the leader from The cooresponding Server running alongside it.