Struct raft::RaftNode

pub struct RaftNode<T: Encodable + Decodable + Send + Clone> {
    // some fields omitted
}

The Raft Distributed Consensus Algorithm requires two RPC calls to be available:

append_entries which is used as both a heartbeat (with no payload) and the primary interface for requests.

request_vote which is used by candidates during campaigns to obtain a vote.

A RaftNode acts as a replicated state machine. The server's role in the cluster depends on it's own status. It will maintain both volatile state (which can be safely lost) and persistent state (which must be carefully stored and kept safe).

Currently, the RaftNode API is not well defined. We are looking for feedback and suggestions.

You can create a cluster like so:

use raft::RaftNode;
use std::old_io::net::ip::SocketAddr;
use std::old_io::net::ip::IpAddr::Ipv4Addr;
// Generally, your nodes will come from a file, or something.
let nodes = vec![
    (0, SocketAddr { ip: Ipv4Addr(127, 0, 0, 1), port: 11110 }),
    (1, SocketAddr { ip: Ipv4Addr(127, 0, 0, 1), port: 11111 }),
    (2, SocketAddr { ip: Ipv4Addr(127, 0, 0, 1), port: 11112 }),
];
// Create the nodes. You recieve a channel back to communicate on.
// TODO: We will probably change this and make it less awkward.
let (command_sender, result_reciever) = RaftNode::<String>::start(
    0,
    nodes.clone(),
    Path::new("/tmp/test0")
);

Note: The Raft paper suggests a minimum cluster size of 3 nodes.

Methods

impl<T: Encodable + Decodable + Debug + Send + 'static + Clone> RaftNode<T>

The implementation of the RaftNode. In most use cases, creating a RaftNode should just be done via ::new().

fn start(id: u64, nodes: Vec<(u64, SocketAddr)>, log_path: Path) -> (Sender<ClientRequest<T>>, Receiver<Result<Vec<(u64, T)>>>)

Creates a new RaftNode with the neighbors specified. id should be a valid index into nodes. The idea is that you can use the same nodes on all of the clients and only vary id.