Struct raft::Server

pub struct Server<L, M> where
    L: Log,
    M: StateMachine,  { /* fields omitted */ }

The Server is responsible for receiving events from peer Server instance or clients, as well as managing election and heartbeat timeouts. When an event is received, it is applied to the local Consensus. The Consensus may optionally return a set of events to be dispatched to either remote peers or clients.

Logging

Server instances log events according to frequency and importance. It is recommended to use at least info level logging when running in production. The warn level is used for unexpected, but recoverable events. The info level is used for infrequent events such as connection resets and election results. The debug level is used for frequent events such as client proposals and heartbeats. The trace level is used for very high frequency debugging output.

Methods

impl<L, M> Server<L, M> where
L: Log,
M: StateMachine,

The implementation of the Server.

fn new(
id: ServerId,
addr: SocketAddr,
store: L,
state_machine: M
) -> ServerBuilder<L, M>

fn run(&mut self) -> Result<()>

Runs a new Raft server in the current thread.

Arguments

id - The ID of the new node.

addr - The address of the new node.

peers - The ID and address of all peers in the Raft cluster.

store - The persistent log store.

state_machine - The client state machine to which client commands will be applied.

fn spawn(
id: ServerId,
addr: SocketAddr,
peers: HashMap<ServerId, SocketAddr>,
store: L,
state_machine: M
) -> Result<JoinHandle<Result<()>>>

Spawns a new Raft server in a background thread.

Arguments

id - The ID of the new node.

addr - The address of the new node.

peers - The ID and address of all peers in the Raft cluster.

store - The persistent log store.

state_machine - The client state machine to which client commands will be applied.