Trait raft::Raft

pub trait Raft<T: Encodable + Decodable + Send + Clone> {
    fn append_entries(term: u64, leader_id: u64, prev_log_index: u64, prev_log_term: u64, entries: Vec<T>, leader_commit: u64) -> (u64, bool);
    fn request_vote(term: u64, candidate_id: u64, last_log_index: u64, last_log_term: u64) -> (u64, bool);
}

The Raft Distributed Consensus Algorithm requires two RPC calls to be available:

append_entries which is used as both a heartbeat (with no payload) and the primary interface for requests.

request_vote which is used by candidates during campaigns to obtain a vote.

Required Methods

fn append_entries(term: u64, leader_id: u64, prev_log_index: u64, prev_log_term: u64, entries: Vec<T>, leader_commit: u64) -> (u64, bool)

Returns (term, success)

fn request_vote(term: u64, candidate_id: u64, last_log_index: u64, last_log_term: u64) -> (u64, bool)

Returns (term, voteGranted)