Module raft::util

This module contains a collection of various tools to use to manipulate and control messages and data associated with raft.

Structs

Union
A convenient struct that handles queries to both HashSet.

Constants

NO_LIMIT
A number to represent that there is no limit.

Functions

is_continuous_ents
Check whether the entry is continuous to the message. i.e msg's next entry index should be equal to the first entries's index

limit_size
Truncates the list of entries down to a specific byte-length of all entries together.

majority
Get the majority number of given nodes count.