Crate raft

Structs

LeaderState
Leader Only Reinitialized after election.

PersistentState
Persistent state Must be updated to stable storage before RPC response.

RaftNode
A RaftNode acts as a replicated state machine. The server's role in the cluster depends on it's own status. It will maintain both volatile state (which can be safely lost) and persistent state (which must be carefully stored and kept safe).

VolatileState
Volatile state

Enums

ClientRequest
Data interchange request format for Client <-> Node Communication. Each variant of this is a command which can be asked of the RaftNode after it is spun up with node.spinup() The node attached to this application will poll it's channel regularly and return results on the channel. If you're wondering where vote requesting is, it's hidden within the module. TODO: Currently requests are not queued or gaurenteed to serviced in order. This is probably a bad thing as most clients will .send() then .recv(). We can probably make a service queue for this.

ClientResponse
Data interchange response format for Client <-> Node communication. Each variant of this is a response to a ClientRequest.

NodeState
Nodes can either be:

RemoteProcedureCall
Data interchange format for RPC calls. These should match directly to the Raft paper's RPC descriptions.

RemoteProcedureResponse
Data interchange format for RPC responses. * Accepted mean that it worked. * Rejected means that rpc.term < node.persistent_state.current_term or if the Node's log doesn't contain the entry at rpc.prev_log_index that maches prev_log_term. The caller should follow the current_leader it is directed to.

Traits

Raft
The Raft Distributed Consensus Algorithm requires two RPC calls to be available: