Crate raft

This is the Raft Distributed Consensus Protocol implemented for Rust. Raft is described as:

Raft is a consensus algorithm that is designed to be easy to understand. It's equivalent to Paxos in fault-tolerance and performance. The difference is that it's decomposed into relatively independent subproblems, and it cleanly addresses all major pieces needed for practical systems.

This implementation utilizes Cap'n Proto for its RPC, mio for it's async event loop.

If this package fails to build for you it is possibly because you do not have the capnp utility installed. You should be able to find appropriate packages for most popular distributions.

Consuming this library

Consuming this library works in a few parts:

Implement or consume a Persistent Log and a State Machine such that they will hook into your application desirably.

Create a Server with those implementations. It will independently fire up and join the cluster.

Interact with the cluster by issuing .propose() and .query() calls via the Client

React to calls to .propose() and .query() from the implemented StateMachine

Persistent Log

A Log represents the replicated, persistent log of your application. It has a strong ordering such that A → B → C and should only act to store information. Entries placed into the log should not be acted on in any way by the consuming application as they have not been committed to the cluster.

Some ideas for a Persistent Log implementation:

A PostgreSQL / SQLite instance.

A plain old file.

A vector in memory (Note: Log compaction is still pending, so be aware of running out!)

It is our belief that in many cases the implementation of Log will be generic to application purposes. You are encouraged to submit your own implementations to us!

State Machine

The StateMachine represents the stateful representation of your application. Events are applied to the StateMachine in the correct ordering at the time of commit. This is where your application should act on information.

In the StateMachine there are both mutable (.apply()) and immutable (.query()) calls. There is a considerable performance difference, as .query() calls do not pass through the durable Log while .apply() events do.

Some ideas for a State Machine implementation:

A Hashmap or key-value store (Example provided)

A single register (Example provided)

Basically anything from std::collections

Client Requests

Client requests are the only way to interact with the Raft cluster. Calls to .propose() and .query() are automatically routed to the relevant Leader node and behave as blocking calls.

This means .propose() won't return until the entry is durably replicated into the log of at least the majority of the cluster and has been commited. .query() will perform better if you wish to only read data and not have it pass through the persisted log.

Reexports

pub use state_machine::StateMachine;

pub use persistent_log::Log;

Modules

messages_capnp

persistent_log
The persistent storage of Raft state.

state_machine
A StateMachine is a single instance of a distributed application. It is the raft libraries responsibility to take commands from the Client and apply them to each StateMachine instance in a globally consistent order.

Structs

Client
The representation of a Client connection to the cluster.

ClientId
The ID of a Raft client.

LogIndex
The index of a log entry.

Server
The Server is responsible for receiving events from peer Server instance or clients, as well as managing election and heartbeat timeouts. When an event is received, it is applied to the local Consensus. The Consensus may optionally return a set of events to be dispatched to either remote peers or clients.

ServerId
The ID of a Raft server. Must be unique among the participants in a consensus group.

Term
The term of a log entry.

Enums

Error
The generic raft::Error is composed of one of the errors that can originate from the various libraries consumed by the library. With the exception of the Raft variant these are generated from try!() macros invoking on io::Error or capnp::Error by using FromError.

RaftError
A Raft Error represents a Raft specific error that consuming code is expected to handle gracefully.

Type Definitions

Result
A simple convienence type.