Enum raft::Error

pub enum Error {
    CapnProto(Error),
    SchemaError(NotInSchema),
    Io(Error),
    Raft(RaftError),
    AddrParse(AddrParseError),
}

The generic raft::Error is composed of one of the errors that can originate from the various libraries consumed by the library. With the exception of the Raft variant these are generated from try!() macros invoking on io::Error or capnp::Error by using FromError.

Variants
CapnProto(Error)

An error originating from the Cap'n Proto library.

SchemaError(NotInSchema)

A specific error produced when a bad Cap'n proto message is discovered.

Io(Error)

Errors originating from std::io.

Raft(RaftError)

Raft specific errors.

AddrParse(AddrParseError)

Errors related to parsing addresses.