Enum raft::Error

pub enum Error {
    CapnProto(Error),
    SchemaError(NotInSchema),
    Io(Error),
    Raft(RaftError),
    AddrParse(AddrParseError),
}

The generic raft::Error is composed of one of the errors that can originate from the various libraries consumed by the library. With the exception of the Raft variant these are generated from try!() macros invoking on io::Error or capnp::Error by using FromError.

Variants

CapnProto

An error originating from the Cap'n Proto library.

SchemaError

A specific error produced when a bad Cap'n proto message is discovered.

Io

Errors originating from std::io.

Raft

Raft specific errors.

AddrParse

Errors related to parsing addresses.