Module raft::persistent_log

The persistent storage of Raft state.

In your consuming application you will want to implement this trait on one of your structures. This could adapt to a database, a file, or even just POD.

Note: Your consuming application should not necessarily interface with this data. It is meant for internal use by the library, we simply chose not to be opinionated about how data is stored.

Structs

MemLog
This is a Log implementation that stores entries in a simple in-memory vector. Other data is stored in a struct. It is chiefly intended for testing.

Enums

Error
Non-instantiable error type for MemLog

Traits

Log
A store of persistent Raft state.