Module raft::raw_node

The raw node of the raft module.

This module contains the value types for the node and it’s connection to other nodes but not the raft consensus itself. Generally, you’ll interact with the RawNode first and use it to access the inner workings of the consensus protocol.

Structs

LightReady
LightReady encapsulates the commit index, committed entries and messages that are ready to be applied or be sent to other peers.

Peer
Represents a Peer node in the cluster.

RawNode
RawNode is a thread-unsafe Node. The methods of this struct correspond to the methods of Node and are described more fully there.

Ready
Ready encapsulates the entries and messages that are ready to read, be saved to stable storage, committed or sent to other peers.

Enums

SnapshotStatus
The status of the snapshot.

Functions

is_empty_snap	Deprecated
For a given snapshot, determine if it’s empty or not.

is_local_msg
Checks if certain message type should be used internally.