Struct raft::RaftNode

pub struct RaftNode<T: Encodable + Decodable + Send + Clone> {
    // some fields omitted
}

A RaftNode acts as a replicated state machine. The server's role in the cluster depends on it's own status. It will maintain both volatile state (which can be safely lost) and persistent state (which must be carefully stored and kept safe).

Methods

impl<T: Encodable + Decodable + Send + Clone> RaftNode<T>

The implementation of the RaftNode. In most use cases, creating a RaftNode should just be done via ::new().

use raft::RaftNode;
use std::old_io::net::ip::SocketAddr;
use std::old_io::net::ip::IpAddr::Ipv4Addr;
use std::collections::HashMap;

let mut nodes = HashMap::new();
nodes.insert(1, SocketAddr { ip: Ipv4Addr(127, 0, 0, 1), port: 11111 });
nodes.insert(2, SocketAddr { ip: Ipv4Addr(127, 0, 0, 1), port: 11112 });
// Create the nodes.
let node = RaftNode::<String>::start(1, nodes.clone());

fn start(id: u64, nodes: HashMap<u64, SocketAddr>) -> (Sender<ClientRequest<T>>, Receiver<Result<Vec<T>, String>>)

Creates a new RaftNode with the neighbors specified. id should be a valid index into nodes. The idea is that you can use the same nodes on all of the clients and only vary id.