Struct raft::Server

pub struct Server<L, M> where L: Log, M: StateMachine {
    // some fields omitted
}

The Server is responsible for receiving events from remote Server or Client instances, as well as setting election and heartbeat timeouts. When an event is received, it is applied to the local Consensus. The Consensus may optionally return a new event which must be dispatched to either the Server or Client which sent the original event, or to all Server instances.

Because messages are passed asyncronously between Server instances, a Server could get into a situation where multiple events are ready to be dispatched to a single remote Server. In this situation, the Server will replace the existing event with the new event, except in one special circumstance: if the new and existing messages are both AppendEntryRequests with the same term, then the new message will be dropped.

Methods

impl<L, M> Server<L, M> where L: Log, M: StateMachine

The implementation of the Server.

fn run(id: ServerId, addr: SocketAddr, peers: HashMap<ServerId, SocketAddr>, store: L, state_machine: M) -> Result<()>

Runs a new Raft server in the current thread.

Arguments

id - The ID of the new node.

addr - The address of the new node.

peers - The ID and address of all peers in the Raft cluster.

store - The persistent log store.

state_machine - The client state machine to which client commands will be applied.

fn spawn(id: ServerId, addr: SocketAddr, peers: HashMap<ServerId, SocketAddr>, store: L, state_machine: M) -> Result<JoinHandle<Result<()>>>

Spawns a new Raft server in a background thread.

Arguments

id - The ID of the new node.

addr - The address of the new node.

peers - The ID and address of all peers in the Raft cluster.

store - The persistent log store.

state_machine - The client state machine to which client commands will be applied.